requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
slack-sdk>=3.30.0
notion-client>=2.2.0
